squash! RevisionFileTreeControl: Ignore ReloadHotkeys
Factor out base class TranslatedControl from GitExtensionsControl
